Update PARA%3A Projects.md to use properties, add uncategorized list
Using properties in the query help narrow down the search results. I have a "Project" property that I use in my "Meetings" blocks, e.g., and all of those blocks were getting picked up in the new "Not Categorized" list. Switching to discriminate on the "type" property fixes this.
 
The "Not Categorized" list is intended to catch projects that I don't properly categorize so I can fix them.